Use consistently int for line numbers
Scintilla uses int not LONG or LONG_PTR for line numbers.

+++ src/TortoiseGitBlame/TortoiseGitBlameData.h
@@ -100,7 +100,7 @@ public:
 		return m_Filenames[line];
 	}
 
-	LONG GetOriginalLineNumber(int line) const
+	int GetOriginalLineNumber(int line) const
 	{
 		return m_OriginalLineNumbers[line];
 	}
